% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loryR.R
\name{renderLoryR}
\alias{renderLoryR}
\title{Widget render function for use in Shiny}
\usage{
renderLoryR(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

