% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{rstudio_gallery}
\alias{rstudio_gallery}
\title{Gallery of RStudio Plots}
\usage{
rstudio_gallery()
}
\description{
rstudio_gallery() is a convenience function to grab all the plots from your RStudio history and
return the images as a \code{list} of \code{base64} encoded images.
}

